+++ packages/sdk-core/src/token-pair.ts
@@ -1,6 +1,7 @@
 import { Token } from '@setheum.js/sdk-core/token';
-import { CurrencyId, TradingPair } from '@setheum.js/types/interfaces';
+import { SetheumPrimitivesTradingPair } from '@polkadot/types/lookup';
 import { assert } from '@polkadot/util';
+import { CombinedCurrencyId } from '.';
 import { AnyApi } from './types';
 
 // class for store token pair
@@ -9,7 +10,7 @@ export class TokenPair {
   private token2: Token;
   private origin: [Token, Token];
 
-  static fromCurrencyId(currency: CurrencyId): TokenPair {
+  static fromCurrencyId(currency: CombinedCurrencyId): TokenPair {
     assert(currency.isDexShare, 'TokenPair.fromCurrencyId should receive CurrencyId which is DexShare');
 
     const [currency1, currency2] = currency.asDexShare;
@@ -17,16 +18,11 @@ export class TokenPair {
     return new TokenPair(Token.fromTokenSymbol(currency1 as any), Token.fromTokenSymbol(currency2 as any));
   }
 
-  static fromCurrencies(currency1: CurrencyId, currency2: CurrencyId): TokenPair {
-    assert(
-      currency1.isToken && currency2.isToken,
-      'TokenPair.fromCurrenciesArray should receive CurrencyId which is TokenSymbol'
-    );
-
+  static fromCurrencies(currency1: CombinedCurrencyId, currency2: CombinedCurrencyId): TokenPair {
     return new TokenPair(Token.fromCurrencyId(currency1), Token.fromCurrencyId(currency2));
   }
 
-  static fromTrandingPair(tradingPair: TradingPair): TokenPair {
+  static fromTrandingPair(tradingPair: SetheumPrimitivesTradingPair): TokenPair {
     const [currency1, currency2] = tradingPair;
 
     return TokenPair.fromCurrencies(currency1, currency2);
@@ -50,15 +46,15 @@ export class TokenPair {
     return [this.token1, this.token2];
   }
 
-  public isEqual(pair: TokenPair, compair?: (token1: Token, token2: Token) => boolean): boolean {
-    return pair.token1.isEqual(this.token1, compair) && pair.token2.isEqual(this.token2, compair);
+  public isEqual(pair: TokenPair, compare?: (token1: Token, token2: Token) => boolean): boolean {
+    return pair.token1.isEqual(this.token1, compare) && pair.token2.isEqual(this.token2, compare);
   }
 
   public toChainData(): [{ Token: string }, { Token: string }] {
     return [this.token1.toChainData() as { Token: string }, this.token2.toChainData() as { Token: string }];
   }
 
-  public toTradingPair(api: AnyApi): TradingPair {
-    return api.createType('TradingPair', this.toChainData());
+  public toTradingPair(api: AnyApi): SetheumPrimitivesTradingPair {
+    return api.createType('SetheumPrimitivesTradingPair', this.toChainData());
   }
 }