+++ packages/sdk-wallet/src/config.ts
@@ -1,1 +1,9 @@
-export const ORACLE_FEEDS_TOKEN = ['SETM', 'SERP', 'DNAR', 'SETR', 'SETUSD'];
+export const ORACLE_FEEDS_TOKEN = [
+  // 'SETM',
+  // 'SERP',
+  // 'DNAR',
+  // 'SETR',
+  // 'SETUSD',
+  // TODO: FIXME - Add cross-chain bridged tokens
+  'sa://0'
+];