+++ packages/types/src/interfaces/cdpEngine/types.ts
@@ -1,15 +1,16 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
+import type { Enum, Option, Struct } from '@polkadot/types-codec';
 import type { Balance } from '@setheum.js/types/interfaces/runtime';
 import type { Rate, Ratio } from '@setheum.js/types/interfaces/support';
-import type { Enum, Option, Struct } from '@polkadot/types';
 
 /** @name ChangeBalance */
 export interface ChangeBalance extends Enum {
   readonly isNoChange: boolean;
   readonly isNewValue: boolean;
   readonly asNewValue: Balance;
+  readonly type: 'NoChange' | 'NewValue';
 }
 
 /** @name ChangeOptionRate */
@@ -17,6 +18,7 @@ export interface ChangeOptionRate extends Enum {
   readonly isNoChange: boolean;
   readonly isNewValue: boolean;
   readonly asNewValue: OptionRate;
+  readonly type: 'NoChange' | 'NewValue';
 }
 
 /** @name ChangeOptionRatio */
@@ -24,12 +26,14 @@ export interface ChangeOptionRatio extends Enum {
   readonly isNoChange: boolean;
   readonly isNewValue: boolean;
   readonly asNewValue: OptionRatio;
+  readonly type: 'NoChange' | 'NewValue';
 }
 
 /** @name LiquidationStrategy */
 export interface LiquidationStrategy extends Enum {
   readonly isAuction: boolean;
   readonly isExchange: boolean;
+  readonly type: 'Auction' | 'Exchange';
 }
 
 /** @name OptionRate */