+++ packages/types/src/interfaces/runtime/types.ts
@@ -1,13 +1,14 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { AuthoritysOriginId, CurrencyId } from '@setheum.js/types/interfaces/primitives';
 import type { Price } from '@open-web3/orml-types/interfaces/traits';
-import type { Bytes, Compact, DoNotConstruct, Enum, GenericAccountId, GenericAccountIndex, GenericBlock, GenericCall, GenericConsensusEngineId, GenericEthereumAccountId, GenericLookupSource, GenericMultiAddress, Int, Null, Option, StorageKey, Struct, U8aFixed, UInt, Vec, u128, u16, u32, u64, u8 } from '@polkadot/types';
+import type { GenericAccountId, GenericAccountIndex, GenericBlock, GenericCall, GenericConsensusEngineId, GenericEthereumAccountId, GenericLookupSource, GenericMultiAddress, StorageKey } from '@polkadot/types';
+import type { Bytes, Compact, DoNotConstruct, Enum, Int, Null, Option, Struct, U8aFixed, UInt, Vec, u128, u16, u32, u64, u8 } from '@polkadot/types-codec';
+import type { ITuple } from '@polkadot/types-codec/types';
 import type { AuthorityId } from '@polkadot/types/interfaces/consensus';
 import type { Signature } from '@polkadot/types/interfaces/extrinsics';
 import type { SystemOrigin } from '@polkadot/types/interfaces/system';
-import type { ITuple } from '@polkadot/types/types';
+import type { AuthoritysOriginId, CurrencyId } from '@setheum.js/types/interfaces/primitives';
 
 /** @name AccountId */
 export interface AccountId extends AccountId32 {}
@@ -73,6 +74,7 @@ export interface ChangesTrieConfiguration extends Struct {
 export interface ChangesTrieSignal extends Enum {
   readonly isNewConfiguration: boolean;
   readonly asNewConfiguration: Option<ChangesTrieConfiguration>;
+  readonly type: 'NewConfiguration';
 }
 
 /** @name CodecHash */
@@ -83,6 +85,13 @@ export interface Consensus extends ITuple<[ConsensusEngineId, Bytes]> {}
 
 /** @name ConsensusEngineId */
 export interface ConsensusEngineId extends GenericConsensusEngineId {}
+
+/** @name CrateVersion */
+export interface CrateVersion extends Struct {
+  readonly major: u16;
+  readonly minor: u8;
+  readonly patch: u8;
+}
 
 /** @name Digest */
 export interface Digest extends Struct {
@@ -108,6 +117,7 @@ export interface DigestItem extends Enum {
   readonly isChangesTrieSignal: boolean;
   readonly asChangesTrieSignal: ChangesTrieSignal;
   readonly isRuntimeEnvironmentUpdated: boolean;
+  readonly type: 'Other' | 'AuthoritiesChange' | 'ChangesTrieRoot' | 'SealV0' | 'Consensus' | 'Seal' | 'PreRuntime' | 'ChangesTrieSignal' | 'RuntimeEnvironmentUpdated';
 }
 
 /** @name EncodedJustification */
@@ -209,8 +219,8 @@ export interface LookupSource extends MultiAddress {}
 /** @name LookupTarget */
 export interface LookupTarget extends AccountId {}
 
-/** @name PalletId */
-export interface PalletId extends LockIdentifier {}
+/** @name ModuleId */
+export interface ModuleId extends LockIdentifier {}
 
 /** @name Moment */
 export interface Moment extends UInt {}
@@ -226,6 +236,7 @@ export interface MultiSigner extends Enum {
   readonly asSr25519: U8aFixed;
   readonly isEcdsa: boolean;
   readonly asEcdsa: U8aFixed;
+  readonly type: 'Ed25519' | 'Sr25519' | 'Ecdsa';
 }
 
 /** @name OpaqueCall */
@@ -244,6 +255,7 @@ export interface Origin extends DoNotConstruct {}
 export interface OriginCaller extends Enum {
   readonly isSystem: boolean;
   readonly asSystem: SystemOrigin;
+  readonly type: 'System';
 }
 
 /** @name PalletId */
@@ -263,6 +275,7 @@ export interface PalletVersion extends Struct {
 export interface Pays extends Enum {
   readonly isYes: boolean;
   readonly isNo: boolean;
+  readonly type: 'Yes' | 'No';
 }
 
 /** @name Perbill */
@@ -297,6 +310,7 @@ export interface ProxyType extends Enum {
   readonly isAuction: boolean;
   readonly isSwap: boolean;
   readonly isLoan: boolean;
+  readonly type: 'Any' | 'CancelProxy' | 'Governance' | 'Auction' | 'Swap' | 'Loan';
 }
 
 /** @name Releases */
@@ -311,6 +325,7 @@ export interface Releases extends Enum {
   readonly isV8: boolean;
   readonly isV9: boolean;
   readonly isV10: boolean;
+  readonly type: 'V1' | 'V2' | 'V3' | 'V4' | 'V5' | 'V6' | 'V7' | 'V8' | 'V9' | 'V10';
 }
 
 /** @name RuntimeDbWeight */